requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
pyyaml>=6.0
click>=8.1.0
pytest>=7.0.0
flask>=3.0.0
